chore: remove marketing surface (CTO PR-1)
First of three CTO spine-consolidation PRs (after Gallery/Playground/Canvas trio #637/#638/#639). Per CTO direction (May 1, 2026): Archmorph is an internal engineering workbench, not a product, has no customers, and ships nothing for sale. Marketing surfaces — landing pages, onboarding intro modals, cookie consent, GDPR DSAR endpoints — make no sense for an internal tool used by platform/cloud engineers.

Removed (frontend):

- LandingPage.jsx (291) + LandingPage.test.jsx (99) — marketing entry page

- CookieBanner.jsx (166) + CookieBanner.test.jsx (72) — GDPR cookie consent

- OnboardingTour.jsx (130) — 2-slide marketing intro modal (ContextualHint extracted to its own file ContextualHint.jsx, kept for spine flow hints in UploadStep/ResultsTable/IaCViewer)

- LegalPages.jsx (302) + LegalPages.test.jsx (82) — terms/privacy/cookie/DSAR consumer pages

Removed (backend):

- routers/legal.py (361) + tests/test_legal_router.py (125) — /api/legal/{terms,privacy,ai-disclaimer,cookies,cookies/consent,data-deletion}

- routers/privacy.py (364) + tests/test_privacy.py (279) — /api/privacy/{legal-documents,consent,dsar,delete,data-practices}

Wiring updated: App.jsx drops three lazy imports + landing/legal route conditionals; App.test.jsx drops three vi.mock blocks; useAppStore.js drops 'landing' and 'legal' from VALID_TABS; Footer.jsx drops the Legal & Privacy button + Shield icon; ChatWidget.jsx drops cookieBannerVisible state + position-shift logic (button is always at bottom-6 now); ContextualHint imports re-pointed; main.py drops privacy/legal_router includes + imports + the v1 mirror entry; routers/platform/__init__.py drops legal/privacy package imports + includes.

Validation: backend pytest 1737 passed/1 skipped/2 xfailed (was 1767, -30 from removed legal+privacy tests); frontend vitest 242 passed/26 files (was 271, -29 from removed Landing/Cookie/Legal tests); OpenAPI contract green; 235 routes mirrored under /api/v1 (was 243, -8 from /api/legal/* and /api/privacy/*).

Net delivered: -3327 LOC across 25 files (vs CTO PR-1 budget of -500/-650 — overshot because PR-1's listed scope explicitly includes both legal.py and privacy.py and their full test suites; also collapses LegalPages.jsx footer button + tab routing). Spine surface reduced by another router pair.

Intentionally kept: ContextualHint primitive (used inside DiagramTranslator spine), Footer + ChatWidget (engineer usability), DisclaimerBanner (separate AI-disclaimer notice not tied to legal/privacy).
