+++ backend/assets/roadmap.json
@@ -408,7 +408,7 @@
     "highlights": [
       "Complete landing page redesign with 9 feature cards",
       "Stats bar (405+ services, 120+ mappings, 3 IaC formats)",
-      "100% free forever \u2014 removed all beta language",
+
       "Buy Me a Coffee link restored in footer",
       "Updated FAQs with accurate feature descriptions",
       "Removed beta preview banner",
