+++ backend/routers/platform/__init__.py
@@ -10,8 +10,6 @@ from routers.health import router as health_router
 from routers.services import router as services_router
 from routers.roadmap import router as roadmap_router
 from routers.feedback import router as feedback_router
-from routers.legal import router as legal_router
-from routers.privacy import router as privacy_router
 from routers.jobs import router as jobs_router
 from routers.compliance_routes import router as compliance_router
 from routers.webhook_routes import router as webhook_routes_router
@@ -27,8 +25,6 @@ domain_router.include_router(health_router)
 domain_router.include_router(services_router)
 domain_router.include_router(roadmap_router)
 domain_router.include_router(feedback_router)
-domain_router.include_router(legal_router)
-domain_router.include_router(privacy_router)
 domain_router.include_router(jobs_router)
 domain_router.include_router(compliance_router)
 domain_router.include_router(webhook_routes_router)
