+++ frontend/src/components/ChatWidget.jsx
@@ -16,22 +16,6 @@ export default function ChatWidget() {
   const [sessionId] = useState(() => `chat-${crypto.randomUUID()}`);
   const messagesEndRef = useRef(null);
   const inputRef = useRef(null);
-  // Offset chat button when cookie banner is visible (#215)
-  const [cookieBannerVisible, setCookieBannerVisible] = useState(false);
-
-  useEffect(() => {
-    // Check cookie consent on mount and after storage changes
-    const checkConsent = () => {
-      const stored = localStorage.getItem('archmorph_cookie_consent');
-      setCookieBannerVisible(!stored);
-    };
-    checkConsent();
-    // Listen for storage changes (banner dismissed)
-    window.addEventListener('storage', checkConsent);
-    // Also re-check periodically (same-window storage writes don't fire 'storage' events)
-    const interval = setInterval(checkConsent, 2000);
-    return () => { window.removeEventListener('storage', checkConsent); clearInterval(interval); };
-  }, []);
 
   useEffect(() => {
     messagesEndRef.current?.scrollIntoView({ behavior: 'smooth' });
@@ -123,7 +107,7 @@ export default function ChatWidget() {
     <>
       <button
         onClick={() => setOpen(!open)}
-        className={`fixed right-6 z-50 w-14 h-14 rounded-full bg-cta hover:bg-cta-hover text-surface shadow-lg shadow-cta/30 flex items-center justify-center transition-all duration-200 cursor-pointer ${cookieBannerVisible ? 'bottom-36' : 'bottom-6'}`}
+        className="fixed right-6 bottom-6 z-50 w-14 h-14 rounded-full bg-cta hover:bg-cta-hover text-surface shadow-lg shadow-cta/30 flex items-center justify-center transition-all duration-200 cursor-pointer"
         aria-label={open ? 'Close chat' : 'Open chat'}
       >
         {open ? <X className="w-6 h-6" /> : <MessageSquare className="w-6 h-6" />}
