--- frontend/src/components/LegalPages.jsx
@@ -1,302 +1,0 @@
-import React from 'react';
-import { Shield, FileText, Brain, Cookie, Scale, ArrowLeft } from 'lucide-react';
-
-const SECTIONS = {
-  terms: {
-    title: 'Terms of Service',
-    icon: Scale,
-    effectiveDate: 'January 15, 2025',
-    content: [
-      {
-        heading: '1. Acceptance of Terms',
-        text: 'By accessing or using Archmorph ("the Service"), you agree to be bound by these Terms of Service. If you do not agree to these terms, you may not use the Service.',
-      },
-      {
-        heading: '2. Description of Service',
-        text: 'Archmorph is an AI-powered cloud architecture translation tool that converts AWS and GCP architecture diagrams into Azure equivalents, generating Infrastructure as Code (Terraform/Bicep), High-Level Design documents, and cost estimates.',
-      },
-      {
-        heading: '3. User Obligations',
-        text: 'You agree to: (a) provide accurate information when creating an account; (b) maintain the security of your credentials; (c) not upload malicious content or attempt to exploit the Service; (d) comply with all applicable laws and regulations; (e) not reverse-engineer or attempt to extract source code from the Service.',
-      },
-      {
-        heading: '4. AI-Generated Output Disclaimer',
-        text: 'All outputs generated by Archmorph, including service mappings, Terraform/Bicep code, cost estimates, and migration roadmaps, are AI-generated recommendations. They should be reviewed by qualified professionals before use in production environments. Archmorph does not guarantee the accuracy, completeness, or fitness of generated outputs for any particular purpose.',
-      },
-      {
-        heading: '5. Intellectual Property',
-        text: 'You retain ownership of all diagrams and architecture designs you upload. Archmorph retains ownership of the Service, its algorithms, and generated templates. Generated IaC code is provided under the MIT License for your use.',
-      },
-      {
-        heading: '6. Free Tool & Vibe-Coding State',
-        text: 'Archmorph is currently a "vibe-coding" project in active development. It is provided completely free of charge, with no subscriptions, billing, or hidden fees. Features ship fast and change often (pardon our dust!), so please expect an evolving experience.',
-      },
-      {
-        heading: '7. Limitation of Liability',
-        text: 'This tool is provided "AS-IS" and "AS-AVAILABLE" without warranties of any kind. As an active build-in-progress passion project, we make no guarantees regarding uptime, stability, or the absolute accuracy of AI mapping outputs. To the maximum extent permitted by law, Archmorph shall not be liable for any indirect, incidental, special, consequential, or punitive damages arising from your use of the Service.',
-      },
-      {
-        heading: '8. Termination',
-        text: 'We may terminate or suspend your access to the Service at any time for violation of these Terms. Upon termination, your right to use the Service ceases immediately. You may request data export before termination takes effect.',
-      },
-      {
-        heading: '9. Changes to Terms',
-        text: 'We may update these Terms from time to time. Material changes will be communicated via email or in-app notification at least 30 days before they take effect. Continued use of the Service after changes constitutes acceptance.',
-      },
-      {
-        heading: '10. Governing Law',
-        text: 'These Terms are governed by the laws of the European Union. Any disputes shall be resolved in the courts of Dublin, Ireland.',
-      },
-    ],
-  },
-  privacy: {
-    title: 'Privacy Policy',
-    icon: Shield,
-    effectiveDate: 'January 15, 2025',
-    content: [
-      {
-        heading: '1. Data Controller',
-        text: 'Archmorph acts as the data controller for personal data processed through the Service. Our Data Protection Officer can be reached at privacy@archmorph.io.',
-      },
-      {
-        heading: '2. Data We Collect',
-        text: 'We collect: (a) Account information (email, name) when you register; (b) Architecture diagrams you upload for analysis; (c) Usage data (features used, session duration, error logs); (d) Technical data (IP address, browser type, device information). We do not collect sensitive personal data or data from children under 16.',
-      },
-      {
-        heading: '3. How We Use Your Data',
-        text: 'We use your data to: (a) Provide the cloud architecture translation service; (b) Process your diagrams through Azure OpenAI for analysis; (c) Improve service accuracy and performance; (d) Communicate service updates and security notices; (e) Comply with legal obligations.',
-      },
-      {
-        heading: '4. Legal Basis for Processing (GDPR Art. 6)',
-        text: 'We process data based on: (a) Consent — for diagram analysis and optional analytics; (b) Contract performance — to deliver the Service you signed up for; (c) Legitimate interest — for service improvement, security monitoring, and fraud prevention.',
-      },
-      {
-        heading: '5. Data Sharing',
-        text: 'We share data only with: (a) Azure OpenAI — as a data processor for AI analysis (under DPA with Microsoft); (b) Azure AD B2C — for authentication; (c) Law enforcement — when legally required. We do not sell your data to third parties.',
-      },
-      {
-        heading: '6. Data Retention',
-        text: 'Uploaded diagrams: Session-only (2-hour TTL, automatically deleted). Account data: Retained while account is active, deleted within 30 days of account closure. Usage analytics: Aggregated and anonymized after 90 days. Audit logs: Retained for 1 year for security compliance.',
-      },
-      {
-        heading: '7. International Data Transfers',
-        text: 'Data is processed within the EU (Azure West Europe and North Europe regions). Where transfers outside the EU are necessary, we rely on Standard Contractual Clauses (SCCs) and the Azure EU Data Boundary.',
-      },
-      {
-        heading: '8. Your Rights (GDPR)',
-        text: 'You have the right to: (a) Access your data (Art. 15); (b) Rectify inaccurate data (Art. 16); (c) Request erasure (Art. 17); (d) Restrict processing (Art. 18); (e) Data portability (Art. 20); (f) Object to processing (Art. 21); (g) Withdraw consent at any time. Submit requests via the Privacy section of your account or email privacy@archmorph.io.',
-      },
-      {
-        heading: '9. Cookies',
-        text: 'We use essential cookies for site functionality and optional analytics cookies. You can manage your cookie preferences via the cookie banner or Settings page. See our Cookie Policy for details.',
-      },
-      {
-        heading: '10. Security',
-        text: 'We implement industry-standard security measures including: TLS 1.3 encryption in transit, AES-256 encryption at rest, SOC 2 Type II compliant infrastructure (Azure), regular security audits, and vulnerability scanning.',
-      },
-      {
-        heading: '11. Contact',
-        text: 'For privacy inquiries: privacy@archmorph.io. For data protection complaints, you may contact your local Data Protection Authority. Our lead supervisory authority is the Irish Data Protection Commission.',
-      },
-    ],
-  },
-  dpa: {
-    title: 'Data Processing Agreement',
-    icon: FileText,
-    effectiveDate: 'January 15, 2025',
-    content: [
-      {
-        heading: '1. Scope',
-        text: 'This Data Processing Agreement ("DPA") supplements our Terms of Service and applies to all processing of personal data by Archmorph on behalf of the Customer, as defined under GDPR Article 28.',
-      },
-      {
-        heading: '2. Categories of Data',
-        text: 'Personal data processed includes: user account details (email, name), architecture diagram content, IP addresses, and usage telemetry. Special category data is not processed.',
-      },
-      {
-        heading: '3. Sub-processors',
-        text: 'Archmorph uses the following sub-processors: (a) Microsoft Azure — cloud infrastructure and AI services; (b) Azure OpenAI Service — diagram analysis (EU-hosted). We will notify customers 30 days before adding new sub-processors.',
-      },
-      {
-        heading: '4. Data Security',
-        text: 'Archmorph implements technical and organizational measures including: encryption at rest and in transit, access controls, audit logging, regular security testing, incident response procedures, and employee training.',
-      },
-      {
-        heading: '5. Data Breach Notification',
-        text: 'In the event of a personal data breach, Archmorph will notify the Customer without undue delay and no later than 72 hours after becoming aware of the breach, providing details as required by GDPR Article 33.',
-      },
-      {
-        heading: '6. Audit Rights',
-        text: 'The Customer has the right to audit Archmorph\'s compliance with this DPA. Audits may be conducted by the Customer or an independent third-party auditor, with reasonable notice and during business hours.',
-      },
-    ],
-  },
-  'ai-disclaimer': {
-    title: 'AI Disclaimer',
-    icon: Brain,
-    effectiveDate: 'January 15, 2025',
-    content: [
-      {
-        heading: '1. AI-Generated Content',
-        text: 'Archmorph uses artificial intelligence (Azure OpenAI) to analyze architecture diagrams and generate cloud migration recommendations, Infrastructure as Code, and cost estimates. All outputs are machine-generated and should be treated as recommendations, not definitive solutions.',
-      },
-      {
-        heading: '2. Accuracy Limitations',
-        text: 'While we strive for high accuracy, AI-generated outputs may contain errors, omissions, or suboptimal recommendations. Service mappings may not reflect the latest Azure service updates. Cost estimates are approximations based on publicly available pricing and may differ from actual costs.',
-      },
-      {
-        heading: '3. Human Review Required',
-        text: 'All AI-generated outputs must be reviewed by qualified cloud architects and engineers before deployment to production environments. Archmorph is not responsible for issues arising from deploying unreviewed AI-generated code or configurations.',
-      },
-      {
-        heading: '4. Data Processing',
-        text: 'Uploaded diagrams are processed through Azure OpenAI\'s API. Microsoft\'s Azure OpenAI data processing terms apply. Your data is not used to train or improve the underlying AI models. Diagram data is processed in the EU region and automatically deleted after your session expires.',
-      },
-      {
-        heading: '5. Continuous Improvement',
-        text: 'We continuously update our service catalog, mapping rules, and AI prompts to improve accuracy. If you encounter inaccurate mappings, please report them through the feedback system to help us improve.',
-      },
-    ],
-  },
-  cookies: {
-    title: 'Cookie Policy',
-    icon: Cookie,
-    effectiveDate: 'January 15, 2025',
-    content: [
-      {
-        heading: '1. What Are Cookies',
-        text: 'Cookies are small text files stored on your device when you visit a website. They help provide a better user experience by remembering preferences and enabling certain features.',
-      },
-      {
-        heading: '2. Cookies We Use',
-        text: 'Necessary cookies: Required for site operation (session management, security tokens, cookie consent preferences). Analytics cookies: Help us understand usage patterns (page views, feature engagement). These are only set with your consent. Functional cookies: Enable enhanced features like theme preferences and language settings.',
-      },
-      {
-        heading: '3. Third-Party Cookies',
-        text: 'We do not use third-party advertising cookies. Analytics data is processed using first-party tools. Azure AD B2C may set authentication-related cookies for sign-in functionality.',
-      },
-      {
-        heading: '4. Managing Cookies',
-        text: 'You can manage your cookie preferences at any time through the cookie banner (click "Cookie Preferences" in the footer). You can also configure your browser to block or delete cookies, though this may affect site functionality.',
-      },
-      {
-        heading: '5. Cookie Retention',
-        text: 'Session cookies are deleted when you close your browser. Persistent cookies (consent preferences) are retained for up to 12 months. You can clear them at any time through your browser settings or our cookie preferences panel.',
-      },
-    ],
-  },
-};
-
-function LegalSection({ section, onBack }) {
-  const data = SECTIONS[section];
-  if (!data) return null;
-  const Icon = data.icon;
-
-  return (
-    <div className="max-w-3xl mx-auto" data-testid={`legal-${section}`}>
-      <button
-        onClick={onBack}
-        className="flex items-center gap-1.5 text-sm text-text-secondary hover:text-cta transition-colors mb-6"
-      >
-        <ArrowLeft className="w-4 h-4" />
-        Back to Legal
-      </button>
-
-      <div className="flex items-center gap-3 mb-6">
-        <div className="w-10 h-10 rounded-xl bg-cta/10 flex items-center justify-center">
-          <Icon className="w-5 h-5 text-cta" />
-        </div>
-        <div>
-          <h2 className="text-xl font-bold text-text-primary">{data.title}</h2>
-          <p className="text-xs text-text-muted">Effective: {data.effectiveDate}</p>
-        </div>
-      </div>
-
-      <div className="space-y-6">
-        {data.content.map((item, i) => (
-          <div key={i} className="bg-secondary/30 rounded-xl p-5 border border-border/50">
-            <h3 className="text-sm font-semibold text-text-primary mb-2">{item.heading}</h3>
-            <p className="text-sm text-text-secondary leading-relaxed">{item.text}</p>
-          </div>
-        ))}
-      </div>
-    </div>
-  );
-}
-
-export default function LegalPages({ onBack }) {
-  const [activeSection, setActiveSection] = React.useState(null);
-
-  if (activeSection) {
-    return <LegalSection section={activeSection} onBack={() => setActiveSection(null)} />;
-  }
-
-  const docs = [
-    { key: 'terms', title: 'Terms of Service', desc: 'Usage terms and conditions', icon: Scale },
-    { key: 'privacy', title: 'Privacy Policy', desc: 'How we handle your data', icon: Shield },
-    { key: 'dpa', title: 'Data Processing Agreement', desc: 'Enterprise data processing terms', icon: FileText },
-    { key: 'ai-disclaimer', title: 'AI Disclaimer', desc: 'AI output limitations and responsibilities', icon: Brain },
-    { key: 'cookies', title: 'Cookie Policy', desc: 'Cookie usage and preferences', icon: Cookie },
-  ];
-
-  return (
-    <div className="max-w-3xl mx-auto" data-testid="legal-pages">
-      {onBack && (
-        <button
-          onClick={onBack}
-          className="flex items-center gap-1.5 text-sm text-text-secondary hover:text-cta transition-colors mb-6"
-        >
-          <ArrowLeft className="w-4 h-4" />
-          Back
-        </button>
-      )}
-
-      <div className="mb-8">
-        <div className="flex items-center gap-3 mb-2">
-          <Shield className="w-6 h-6 text-cta" />
-          <h1 className="text-2xl font-bold text-text-primary">Legal & Privacy</h1>
-        </div>
-        <p className="text-sm text-text-muted">
-          Transparency and compliance are core to Archmorph. Review our legal documents below.
-        </p>
-      </div>
-
-      <div className="grid gap-3">
-        {docs.map(doc => {
-          const Icon = doc.icon;
-          return (
-            <button
-              key={doc.key}
-              onClick={() => setActiveSection(doc.key)}
-              className="flex items-center gap-4 p-4 bg-secondary/30 rounded-xl border border-border/50 hover:border-cta/30 hover:bg-secondary/50 transition-all text-left group"
-              data-testid={`legal-link-${doc.key}`}
-            >
-              <div className="w-9 h-9 rounded-lg bg-cta/10 flex items-center justify-center shrink-0 group-hover:bg-cta/20 transition-colors">
-                <Icon className="w-4 h-4 text-cta" />
-              </div>
-              <div>
-                <p className="text-sm font-medium text-text-primary group-hover:text-cta transition-colors">{doc.title}</p>
-                <p className="text-xs text-text-muted">{doc.desc}</p>
-              </div>
-            </button>
-          );
-        })}
-      </div>
-
-      <div className="mt-8 p-4 bg-cta/5 rounded-xl border border-cta/10">
-        <h3 className="text-sm font-medium text-text-primary mb-1">Your Privacy Rights</h3>
-        <p className="text-xs text-text-muted mb-3">
-          Under GDPR, you have the right to access, export, or delete your data at any time.
-        </p>
-        <div className="flex gap-2">
-          <a
-            href="mailto:privacy@archmorph.io"
-            className="px-3 py-1.5 text-xs font-medium text-cta border border-cta/30 rounded-lg hover:bg-cta/10 transition-colors"
-          >
-            Contact DPO
-          </a>
-        </div>
-      </div>
-    </div>
-  );
-}
